API

Documentation for MixedModelsBLB.jl's functions.

Index

MixedModelsBLB.SubsetEstimates

MixedModelsBLB.blbEstimates

MixedModelsBLB.blb_db

MixedModelsBLB.blb_full_data

MixedModelsBLB.blb_one_subset

MixedModelsBLB.confint

Functions

MixedModelsBLB.blb_one_subset — Function

blb_one_subset(rng, m; N, subset_size, n_boots, method, solver, verbose, nonparametric_boot)

Performs Bag of Little Bootstraps on a subset.

Positional arguments

rng: random number generator. Default to the global rng.

m: an object of type blblmmModel or WSVarLmmModel

Keyword arguments

N: number of individuals/clusters in the full dataset.

subset_size: number of individuals/clusters in each subset.

n_boots: number of bootstrap iterations. Default to 1000.

method: a symbol, either :ML or :WiSER.

solver: solver for the optimization problem.

verbose: Bool, whether to print bootstrap progress.

nonparametric_boot: Bool, whether to use nonparametric bootstrap. For WiSER models, only nonparametric bootstrap is supported.

Values

subset_estimates: an object of type SubsetEstimates

MixedModelsBLB.blb_full_data — Function

blb_full_data(rng, datatable; feformula, reformula, wsvarformula, id_name, cat_names, subset_size, n_subsets, n_boots, method, solver, verbose,  nonparametric_boot)

Performs Bag of Little Bootstraps on the full dataset

Positional arguments

rng: random number generator. Default to the global rng.

datatable: a data table type that is compatible with Tables.jl.

Keyword arguments

feformula: model formula for the fixed effects.

reformula: model formula for the random effects.

wsvarformula: model formula for the fixed effects of the within-subject variance. For linear mixed models, it should be @formula(y ~ 1). Only need to be specified when method = :WiSER.

id_name: name of the cluster identifier variable. String.

cat_names: a vector of the names of the categorical variables.

subset_size: number of clusters in the subset.

n_subsets: number of subsets.

n_boots: number of bootstrap iterations. Default to 1000

solver: solver for the optimization problem.

method: fitting the model by maximum-likelihood (:ML) or GEE (:WiSER)

verbose: Bool, whether to print bootstrap progress (percentage completion)

nonparametric_boot: Bool, whether to use Nonparametric bootstrap

Values

result: an object of the blbEstimates type

MixedModelsBLB.blb_db — Function

blb_db(rng, con, table_name; feformula, reformula, id_name, cat_names, subset_size, n_subsets, n_boots, solver, verbose,  nonparametric_boot)

Performs Bag of Little Bootstraps on databases.

Positional arguments

rng: random number generator. Default to the global rng.

con: an object of type MySQL.Connection created by the function DBInterface.connect.

table_name: table name for the longitudinal data.

Keyword arguments

feformula: model formula for the fixed effects.

reformula: model formula for the random effects.

id_name: name of the cluster identifier variable. String.

cat_names: a vector of the names of the categorical variables.

subset_size: number of clusters in the subset.

n_subsets: number of subsets.

n_boots: number of bootstrap iterations. Default to 1000

solver: solver for the optimization problem.

verbose: Bool, whether to print bootstrap progress (percentage completion)

nonparametric_boot: Bool, whether to use Nonparametric bootstrap

Values

result: an object of the blbEstimates type

MixedModelsBLB.SubsetEstimates — Type

SubsetEstimates

BLB linear mixed model estimates from one subset.

MixedModelsBLB.blbEstimates — Type

blbEstimates

BLB linear mixed model estimates, which contains blb parameters and a vector of SubsetEstimates.

MixedModelsBLB.confint — Function

confint(subset_ests, level)

Calculate confidence intervals using estimates from one subset.

Positional arguments

subset_ests: an object of type SubsetEstimates

level: confidence level, usually set to 0.95

confint(blb_ests, level)

Calculate confidence intervals using estimates from all subsets.

Positional arguments

blb_ests: an object of type blbEstimates

level: confidence level, usually set to 0.95

Missing docstring.

Missing docstring for fixef. Check Documenter's build log for details.

Missing docstring.

Missing docstring for vc. Check Documenter's build log for details.